\name{BTSPAS-internal}
\alias{BTSPAS-internal}
\alias{PredictivePosterior.TSPDE}
\alias{PredictivePosterior.TSPDE.WHCH}
\alias{PredictivePosterior.TSPDE.WHCH2}
\alias{PredictivePosterior.TSPDE.WHSteel}
\alias{PredictivePosterior.TSPNDE}
\alias{PredictivePosteriorPlot.TSPDE}
\alias{PredictivePosteriorPlot.TSPDE.WHCH}
\alias{PredictivePosteriorPlot.TSPDE.WHCH2}
\alias{PredictivePosteriorPlot.TSPDE.WHSteel}
\alias{PredictivePosteriorPlot.TSPNDE}
\alias{RunTime}
\alias{SimplePetersen}
\alias{TestIfPool}
\alias{zzz}

\title{ Internal functions not normally called by the user}
\description{
These are internal functions not normally called by the user.
(a) Compute the Predictive Posterior Distributions and make the
Bayesian p-value plots.
(b) Estimate the runtimings once the posterior distribution is found.
(c) Compute simple Petersen estimates
(d) Miscellaneous functions
}
