% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{expit}
\title{Logit and anti-logit function.}
\usage{
logit(p)

expit(theta)
}
\arguments{
\item{p}{probability between 0 and 1.}

\item{theta}{logit between -infinity and +infinity}
}
\value{
Computed logit or anti-logit
}
\description{
Compute the logit or anti-logit.
}
\examples{

##---- compute the logit and its inverse
logitp <- logit(.3)
p <- expit(-.84)

}
\author{
C.J.Schwarz \email{cschwarz@stat.sfu.ca}
}
\keyword{~misc}
