% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Create an acf plot of a parameter}
\usage{
plot_acf(mcmc.sample)
}
\arguments{
\item{mcmc.sample}{Data frame with 2 columns. parm, and sample. A separate ACF plot is generated for each parameter using facet_wrap}
}
\value{
acf plot(s) as an ggplot2 object
}
\description{
Create an acf plot of a parameter
}
\keyword{internal}
