% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Create posterior plot of a parameter with credible interval limits shown as vertical lines}
\usage{
plot_posterior(mcmc.sample, alpha = 0.05, ncol = 1)
}
\arguments{
\item{mcmc.sample}{Data frame with 2 columns. parm, and sample. A separate posterior plot is generated for each parameter.}

\item{alpha}{Used to determine credible interval.}

\item{ncol}{Number of columns in the plot (default=1).}
}
\value{
Posterior plot(s) as an ggplot2 object
}
\description{
Create posterior plot of a parameter with credible interval limits shown as vertical lines
}
\keyword{internal}
