% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTime.R
\name{RunTime}
\alias{RunTime}
\title{Compute percentiles of the run timing distribution.}
\usage{
RunTime(time, U, prob = seq(0, 1, 0.1))
}
\arguments{
\item{time}{List of sample weeks. It is assumed that there are no salmon prior to the first value}

\item{U}{matrix of posterior samples. Each row is a sample from the posterior.}

\item{prob}{Quantiles of the run timing to estimate.}
}
\value{
An MCMC object with samples from the posterior distribution. A
series of graphs and text file are also created in the working directory.
}
\description{
Take the posterior sample of U[1,...nstrata] and compute the percentiles of the run
timing. 
This uses the quantile() function from the "actuar" package which is designed to compute
quantiles of grouped data.
}
\references{
Refer to the Trinity River Restoration Project report by
Schwarz, C.J. et al. (2009) available at
\url{http://www.stat.sfu.ca/~cschwarz/Consulting/Trinity/Phase2}. Please
contact \email{cschwarz@stat.sfu.ca} for more details. %% ~put references to
the literature/web site here ~
}
\author{
Bonner, S.J. \email{s.bonner@stat.ubc.ca} and Schwarz, C. J.
\email{cschwarz@stat.sfu.ca}
}
