\name{BTSPAS-internal}
\alias{BTSPAS-internal}
\alias{genInitVals}
\alias{genInitValsChain}
\alias{genInitsTTln}
\alias{genInitsTTnp}
\alias{lnTheta}
\alias{plot_logitP}
\alias{PredictivePosterior.TSPDE}
\alias{PredictivePosterior.TSPDE.WHCH}
\alias{PredictivePosterior.TSPDE.WHCH2}
\alias{PredictivePosterior.TSPDE.WHSteel}
\alias{PredictivePosterior.TSPNDE}
\alias{PredictivePosterior.TSPNDENP}
\alias{PredictivePosteriorPlot.TSPDE}
\alias{PredictivePosteriorPlot.TSPDE.WHCH}
\alias{PredictivePosteriorPlot.TSPDE.WHCH2}
\alias{PredictivePosteriorPlot.TSPDE.WHSteel}
\alias{PredictivePosteriorPlot.TSPNDE}
\alias{PredictivePosterior.TSPNDENP}
\alias{RunTime}
\alias{run.openbugs}
\alias{run.jags}
\alias{run.MCMC}
\alias{simTSPNDE}
\alias{SimplePetersen}
\alias{TestIfPool}
\alias{visualize.muTT.prior}  % visualize what a prior for muTT looks like
\alias{make.muTT.prior}       % convert a prior for muTT to the internal form needed
\alias{zzz}

\title{ Internal functions not normally called by the user}
\description{
These are internal functions not normally called by the user.
(a) Compute the Predictive Posterior Distributions and make the
Bayesian p-value plots.
(b) Estimate the runtimings once the posterior distribution is found.
(c) Compute simple Petersen estimates.
(d) Generate initial values.
(e) Miscellaneous functions.
}
