\name{BTSPAS-internal}
\alias{BTSPAS-internal}
\alias{genInitVals}
\alias{genInitValsChain}
\alias{genInitsTTnp}
\alias{lnTheta}
\alias{plot_logitP}
\alias{run.openbugs}
\alias{run.jags}
\alias{run.MCMC}
\alias{simTSPNDE}
\alias{visualize.muTT.prior}  % visualize what a prior for muTT looks like
\alias{make.muTT.prior}       % convert a prior for muTT to the internal form needed

\title{ Internal functions not normally called by the user}
\description{
These are internal functions not normally called by the user.
(a) Compute the Predictive Posterior Distributions and make the
Bayesian p-value plots.
(b) Estimate the runtimings once the posterior distribution is found.
(c) Compute simple Petersen estimates.
(d) Generate initial values.
(e) Miscellaneous functions.
}
