% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTime.R
\name{RunTime}
\alias{RunTime}
\title{Compute percentiles of the run timing distribution.}
\usage{
RunTime(time, U, prob = seq(0, 1, 0.1))
}
\arguments{
\item{time}{List of sample weeks. It is assumed that there are no salmon prior to the first value}

\item{U}{matrix of posterior samples. Each row is a sample from the posterior.}

\item{prob}{Quantiles of the run timing to estimate.}
}
\value{
An MCMC object with samples from the posterior distribution. A
series of graphs and text file are also created in the working directory.
This information is now added to the fit object as well and so it is unlikely
that you will use this function.
}
\description{
Take the posterior sample of U[1,...nstrata] and compute the percentiles of the run
timing. 
This uses the quantile() function from the "actuar" package which is designed to compute
quantiles of grouped data.
}
\references{
Bonner, S. J., & Schwarz, C. J. (2011). 
Smoothing population size estimates for Time-Stratified Mark-Recapture experiments Using Bayesian P-Splines. 
Biometrics, 67, 1498-1507. 
\url{https://doi.org/10.1111/j.1541-0420.2011.01599.x}

Schwarz, C. J., & Dempson, J. B. (1994).
Mark-recapture estimation of a salmon smolt population. 
Biometrics, 50, 98-108.

Schwarz, C.J. et al. (2009) Trinity River Restoration Project Report available at
\url{http://www.stat.sfu.ca/~cschwarz/Consulting/Trinity/Phase2}
}
\author{
Bonner, S.J. \email{sbonner6@uwo.ca} and 
Schwarz, C. J. \email{cschwarz.stat.sfu.ca@gmail.com}.
}
